match contracts with compound v2

+++ contracts/ComptrollerStorage.sol
@@ -6,27 +6,28 @@ import "./PriceOracle.sol";
 
 contract UnitrollerAdminStorage {
     /**
-     * @notice Administrator for this contract
-     */
+    * @notice Administrator for this contract
+    */
     address public admin;
 
     /**
-     * @notice Pending administrator for this contract
-     */
+    * @notice Pending administrator for this contract
+    */
     address public pendingAdmin;
 
     /**
-     * @notice Active brains of Unitroller
-     */
+    * @notice Active brains of Unitroller
+    */
     address public comptrollerImplementation;
 
     /**
-     * @notice Pending brains of Unitroller
-     */
+    * @notice Pending brains of Unitroller
+    */
     address public pendingComptrollerImplementation;
 }
 
 contract ComptrollerV1Storage is UnitrollerAdminStorage {
+
     /**
      * @notice Oracle which gives the price of any given asset
      */
@@ -35,34 +36,38 @@ contract ComptrollerV1Storage is UnitrollerAdminStorage {
     /**
      * @notice Multiplier used to calculate the maximum repayAmount when liquidating a borrow
      */
-    uint256 public closeFactorMantissa;
+    uint public closeFactorMantissa;
 
     /**
      * @notice Multiplier representing the discount on collateral that a liquidator receives
      */
-    uint256 public liquidationIncentiveMantissa;
+    uint public liquidationIncentiveMantissa;
 
     /**
      * @notice Max number of assets a single account can participate in (borrow or use as collateral)
      */
-    uint256 public maxAssets;
+    uint public maxAssets;
 
     /**
      * @notice Per-account mapping of "assets you are in", capped by maxAssets
      */
     mapping(address => CToken[]) public accountAssets;
+
 }
 
 contract ComptrollerV2Storage is ComptrollerV1Storage {
     struct Market {
         // Whether or not this market is listed
         bool isListed;
+
         //  Multiplier representing the most one can borrow against their collateral in this market.
         //  For instance, 0.9 to allow borrowing 90% of collateral value.
         //  Must be between 0 and 1, and stored as a mantissa.
-        uint256 collateralFactorMantissa;
+        uint collateralFactorMantissa;
+
         // Per-market mapping of "accounts in this asset"
         mapping(address => bool) accountMembership;
+
         // Whether or not this market receives COMP
         bool isComped;
     }
@@ -72,6 +77,7 @@ contract ComptrollerV2Storage is ComptrollerV1Storage {
      * @dev Used e.g. to determine if a market is supported
      */
     mapping(address => Market) public markets;
+
 
     /**
      * @notice The Pause Guardian can pause certain actions as a safety mechanism.
@@ -91,6 +97,7 @@ contract ComptrollerV3Storage is ComptrollerV2Storage {
     struct CompMarketState {
         // The market's last updated compBorrowIndex or compSupplyIndex
         uint224 index;
+
         // The block number the index was last updated at
         uint32 block;
     }
@@ -99,10 +106,10 @@ contract ComptrollerV3Storage is ComptrollerV2Storage {
     CToken[] public allMarkets;
 
     /// @notice The rate at which the flywheel distributes COMP, per block
-    uint256 public compRate;
+    uint public compRate;
 
     /// @notice The portion of compRate that each market currently receives
-    mapping(address => uint256) public compSpeeds;
+    mapping(address => uint) public compSpeeds;
 
     /// @notice The COMP market supply state for each market
     mapping(address => CompMarketState) public compSupplyState;
@@ -111,13 +118,13 @@ contract ComptrollerV3Storage is ComptrollerV2Storage {
     mapping(address => CompMarketState) public compBorrowState;
 
     /// @notice The COMP borrow index for each market for each supplier as of the last time they accrued COMP
-    mapping(address => mapping(address => uint256)) public compSupplierIndex;
+    mapping(address => mapping(address => uint)) public compSupplierIndex;
 
     /// @notice The COMP borrow index for each market for each borrower as of the last time they accrued COMP
-    mapping(address => mapping(address => uint256)) public compBorrowerIndex;
+    mapping(address => mapping(address => uint)) public compBorrowerIndex;
 
     /// @notice The COMP accrued but not yet transferred to each user
-    mapping(address => uint256) public compAccrued;
+    mapping(address => uint) public compAccrued;
 }
 
 contract ComptrollerV4Storage is ComptrollerV3Storage {
@@ -125,23 +132,23 @@ contract ComptrollerV4Storage is ComptrollerV3Storage {
     address public borrowCapGuardian;
 
     // @notice Borrow caps enforced by borrowAllowed for each cToken address. Defaults to zero which corresponds to unlimited borrowing.
-    mapping(address => uint256) public borrowCaps;
+    mapping(address => uint) public borrowCaps;
 }
 
 contract ComptrollerV5Storage is ComptrollerV4Storage {
     /// @notice The portion of COMP that each contributor receives per block
-    mapping(address => uint256) public compContributorSpeeds;
+    mapping(address => uint) public compContributorSpeeds;
 
     /// @notice Last block at which a contributor's COMP rewards have been allocated
-    mapping(address => uint256) public lastContributorBlock;
+    mapping(address => uint) public lastContributorBlock;
 }
 
 contract ComptrollerV6Storage is ComptrollerV5Storage {
     /// @notice The rate at which comp is distributed to the corresponding borrow market (per block)
-    mapping(address => uint256) public compBorrowSpeeds;
+    mapping(address => uint) public compBorrowSpeeds;
 
     /// @notice The rate at which comp is distributed to the corresponding supply market (per block)
-    mapping(address => uint256) public compSupplySpeeds;
+    mapping(address => uint) public compSupplySpeeds;
 }
 
 contract ComptrollerV7Storage is ComptrollerV6Storage {
@@ -149,5 +156,5 @@ contract ComptrollerV7Storage is ComptrollerV6Storage {
     bool public proposal65FixExecuted;
 
     /// @notice Accounting storage mapping account addresses to how much COMP they owe the protocol.
-    mapping(address => uint256) public compReceivable;
+    mapping(address => uint) public compReceivable;
 }
